Agrega al model de Cliente y Pedido la relacion con Ubicacion.

+++ src/main/java/sic/modelo/dto/UbicacionDTO.java
@@ -12,9 +12,9 @@ public class UbicacionDTO {
 
   private String descripcion;
 
-  private Long latitud;
+  private Double latitud;
 
-  private Long longitud;
+  private Double longitud;
 
   private String calle;
